+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-06-20
+
 ### Added
 
 - Add support for unified cluster-azure app. With cluster-azure v0.14.0 and newer, default apps are deployed with cluster-azure nd default-apps-azure app is not deployed anymore.
@@ -389,7 +391,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/giantswarm/clustertest/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/giantswarm/clustertest/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/giantswarm/clustertest/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/giantswarm/clustertest/compare/v1.0.0...v1.0.1